adding durin data to tritonBench
Summary:
Adding durin data that is pulled from pastry for now.

Additionally for times that we are unable to run specific shapes on a given kernel we do not count those metrics into the aggregates.

Without this any kernel that does not support a wide range of shapes will suddenly have much better latency performance due to the static number of rows used as the denomenator of the average so far

Next steps:
1. extend support for the rest of the operators
2. get a new cut of shapes that are more supported in tritonBench OR figure out exactly why the current shapes are not all supported in each kernel

Reviewed By: FindHao

Differential Revision: D65607643

fbshipit-source-id: 270b46aea8ca0ec935fb23696b995bda1fc81483