Add ufmt linter for pyproject (#47)
Summary:
Add linter to make sure PR is consistent with internal `arc lint`.

To format the code, run the following in the repo directory:

```
ufmt format .
```

We also test 25 operators in our H100 CI. Note that many of the flash_attention backends do not work right now and we need to fix them.

Pull Request resolved: https://github.com/pytorch-labs/tritonbench/pull/47

Reviewed By: FindHao

Differential Revision: D65709181

Pulled By: xuzhao9

fbshipit-source-id: 5b013906e7b04c8ee41d74db5756de08eec5b5b2