add conditions to determine triton ws
Summary:
This is to improve the ottb UX of the flash_attention operator. Right now it will break because warp spec has not yet upstreamed to main. Add predicates to find out if warp spec is available.

Plus, we only enable ThunderKittens in OSS as the fbcode version is outdated.

Reviewed By: FindHao

Differential Revision: D66402556

fbshipit-source-id: f15b079d747e96af87cff0a7ce22d41e63d62843