enabling fp8 gemm on amd
Summary:
currently we were trying to import cublas which was getting a `RuntimeError` instead of the moduleNotFound error. adding the exception to the `try except` clause with logged statements

This was caught originally via CI additions we are doing in the next diff for AMD integration. see https://www.internalfb.com/intern/testinfra/diagnostics/4785074862901046.844425121438624.1733874779/ for the failure that caused this diff

Reviewed By: nmacchioni

Differential Revision: D67065556

fbshipit-source-id: ede7042bc85a0d5d487f33d281cb8ab09cff135e