Fix GEMM for AMD
Summary:
I was originally trying to run tritonbench in OSS, but there was a missing is_hip() check when not in fbcode.

The reason for this first change is_hip() check change change is that if you are building with a custom version of PyTorch the version string is something that resembles "2.6.0a0+gitf0f47b7".

This ended up cascading into a "fix tritonbench GEMM for AMD"-thing. A bunch of various small changes required (e.g., when skipping tests also skip tests that are disabled due to missing CUDA/HIP libraries)

Reviewed By: xuzhao9

Differential Revision: D67803458

fbshipit-source-id: 1065e722f8f89a4bf741f0c3662fd3d661eecbf8