More aggressively tune the GPU and benchmark (#143)
Summary:
Through the OSS metrics tracking we found the day-to-day run variance is larger than we expected (~4%) where we are shooting for 2% max variance.

This PR will tune the GPU more aggressively and run gemm and softmax benchmarks with cudagraph to see if metrics can be more stabilized.

Pull Request resolved: https://github.com/pytorch-labs/tritonbench/pull/143

Test Plan: CI

Reviewed By: adamomainz

Differential Revision: D68644171

Pulled By: xuzhao9

fbshipit-source-id: ea3b34836da536719176d36d1b301f048d8038cd