Disable TMA GEMM Kernels for AMD machines
Summary:
When running TritonBench for the GEMM operator without --only, the system will attempt to pickup every benchmark kernel. This crashes the benchmark when run with an AMD GPU because the TMA kernels are selected, but those aren't supported for AMD.

This PR disables those kernels by checking that supports_tma().

Reviewed By: xuzhao9

Differential Revision: D69094403

fbshipit-source-id: 62819575f65f3099d55a35874162d102b538b801