adding default logging group for tritonbench
Summary: currently if a user wants to log data but does not set the logging group then the field is set to null. this creates some oddities in our data. Setting a very reasonable default here to use the operator name as the logging group (which is what it is set to a lot of the time anyway but not always)

Reviewed By: karthik-man, danzimm, nmacchioni

Differential Revision: D69164555

fbshipit-source-id: 883e34e8f94f9aa2890b5fc90bc888dc7f7c2248