Add DecomposeK matmul kernel
Summary:
Add decomposeK to TritonBench based off of D63917874. Split the K dimension for very large K with small M, N to optimize matmul by decomposing the ```aten.mm``` into a ```aten.bmm```. This approach is also friendly to epilogue fusion in contrast to adding **SPLIT_K** to the kernel itself, as in ```triton_ops_matmul```.

We see almost similar performance to ```aten.mm``` with kPartitions=256 and kPartitions=512 for SPLIT_K shapes on TritonBench. Accuracy does not seem to be affected with this approach.

However, we are currently unable to autotune the kPartitions parameter, which we want to make this the most performant, similar issue: https://docs.google.com/document/d/1K1DwmVkzqoB_uDoWkPvmOa2xra5yko5JmQzy5Jxy-rg/edit?disco=AAABgdS0OKA

Reviewed By: sijiac

Differential Revision: D71368869

fbshipit-source-id: 424e39b4971faae61f2789adb8f1f2240f08abeb