Rename splitk kernel from _kernel
Summary:
When collecting traces with ATT you must provide the name of a kernel upon which to match. When the kernel name is `_kernel` you are likely to match on the first kernel (since most kernels contain _kernel in the name).

Changing to _splitk_kernel enables actually capture the correct kernel of interest.

Reviewed By: xuzhao9

Differential Revision: D71907011

fbshipit-source-id: f93fabd427ecfa32b038c623b89ca5fc1d9a6ac6