Append Triton type to Scuba table logging (#198)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/tritonbench/pull/198

Adds the logic to include a new column, `triton_type` to the scuba table for benchmark logging. This column will be used to track which version of Triton is used in experiments/testing. Currenlty this will be stable and trunk, but in the future may include the trunk + semantic patch version.

Reviewed By: xuzhao9

Differential Revision: D72975732