Add OAI's warp spec implementation to TritonBench (#242)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/tritonbench/pull/242

Adds the implementation of OAI's updated tutorial reduced to just warp spec in TritonBench. A followup PR will update this tutorial to have proper proton support (hence the filename).

A followup diff will integrate the logic for doing Proton profiling. The reasoning for splitting this into a separate file is that this warp spec changes seems to depend on Triton Trunk (not Beta). As a result we will need to use the Proton APIs in OSS and cannot depend on Beta's. Having both in the same file seemed too confusing for me.

Reviewed By: manman-ren

Differential Revision: D76376028