Fix jagged_sum OOM and adjust default values
Summary: `jagged_to_nested_tensor` stores all Tensors upfront, which quickly leads to OOMs. By yielding them one at a time instead they can be garbage collected as runs complete. I also noticed that it is possible to substantially improve the nested tensor generation by allocating on large Tensor and slicing it, which means we only need to launch one kernel (since slicing is a host-only operation), rather than launching B randn kernels. (B is on the order of tens or hundreds of thousands.) And finally, I pared back the defaults since the standard behavior was a massive sweep, but the OOM bug tells me it was never actually run. (Which means it should be fine to go with a more conservative set.)

Reviewed By: xuzhao9

Differential Revision: D76911972

fbshipit-source-id: 57b0201a2ead4526f27bdc15ef60590016e7e16f