Remove unused os import in low_mem_dropout operator
Summary:
## Instructions about RACER Diffs:
You will receive **FULL CREDIT** (ETS) for this diff if you either:
 - [**Recommended**]Commandeer and land this diff

Once landed, feel free to claim the associated task and the EYS project.

 This diff is removing the unused imports from Code Quality Score from [Monetization codehub](https://fburl.com/quality/58ozwacb):
 - Accept and ship the diff (racer do not do auto land)
 - [**Recommended**]Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generage.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T229209355. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [RACER Feedback and Q&A](https://fb.workplace.com/groups/2477474979269093) group.
## Summary:
Removed an unused 'os' import from the low_mem_dropout operator.py file as identified by the flake8 rule F401.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=a8813953-53e3-11f0-8aa9-9299bdf9c8b5&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=a8813953-53e3-11f0-8aa9-9299bdf9c8b5&tab=Trace)

Reviewed By: FindHao

Differential Revision: D77481109

fbshipit-source-id: a2eefde307c14e67b9ac2242fae37357cb0f526f